chore: cloud editing e2e tests
