+++ web-admin/src/routes/[organization]/-/create-project/+page.svelte
@@ -63,7 +63,7 @@
     {:else}
       <RillLogoSquareNegative size="36px" />
       <div class="text-2xl font-extrabold text-fg-accent text-center">
-        Create {hasProjects ? "your first" : "a new"} project
+        Create {hasProjects ? "a new" : "your first"} project
       </div>
 
       <div
@@ -71,7 +71,7 @@
       >
         <div>
           <div class="text-base font-semibold">
-            Name your {hasProjects ? "first " : ""}project
+            Name your {hasProjects ? "" : "first "}project
           </div>
           <div class="text-sm text-fg-muted">
             You can rename anytime from project settings.
