+++ web-local/tests/utils/dataSpecifcHelpers.ts
@@ -6,7 +6,8 @@ import {
   wrapRetryAssertion,
 } from "./commonHelpers";
 import { createModel } from "./modelHelpers";
-import { createSourceV2 } from "./sourceHelpers";
+
+import { createLocalFileSource } from "@rilldata/web-common/tests/utils/source-helpers.ts";
 
 export interface TimeSeriesValue {
   ts: string;
@@ -25,7 +26,7 @@ export const AD_BIDS_EXPLORE_PATH =
 export async function createAdBidsModel(page: Page) {
   await Promise.all([
     waitForProfiling(page, "AdBids", ["publisher", "domain", "timestamp"]),
-    createSourceV2(page, "AdBids.csv", "/models/AdBids.yaml"),
+    createLocalFileSource(page, "AdBids.csv", "/models/AdBids.yaml"),
   ]);
 
   await createModel(page, "AdBids_model.sql");
